python/pyproject.toml
[tool.poetry]
authors = ["Bucketeer team <bucketeer@users.noreply.github.com>"]
description = ""
name = "python"
version = "0.1.0"

[tool.poetry.dependencies]
APScheduler = "3.7.0"
PyMySQL = "1.0.2"
pandas = "^1.2.5"
prometheus-client = "0.7.1"
pystan = "2.19.1.1"
python = "3.7.6"
scipy = "^1.5.3"
environs = "^9.3.4"
grpc-health-checking = "0.0.1"
grpcio-health-checking = "1.27.2"
google-cloud-logging = "1.15.1"
prometheus-async = "19.2.0"
aiohttp = "3.7.4.post0"
google-cloud = "^0.34.0"
protobuf = "^3.20.1"
google-api-python-client = "^2.45.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pycodestyle = "^2.7.0"
pytest = "^6.2.5"
pytest-mock = "^3.7.0"
grpcio-tools = "^1.45.0"
black = "^22.3.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
target_version = ['py37']